+++ examples/matrix/matmul-host/src/lib.rs
@@ -69,7 +69,7 @@ pub fn run_host_matmul<'ctx>(
     println!("GPU execution time: {:?}", elapsed);
 
     let config = gpu_host::gpu_config!(grid_dim, grid_dim, 1, block_dim, block_dim, 1, 0);
-    let d_c_c = gpu::GlobalThreadChunk::new_from_host(&mut d_c, gpu::Map2D::new(n as u32));
+    let d_c_c = gpu::GlobalThreadChunk::new_from_host(&mut d_c, gpu::Map2D::new(n));
     let start = std::time::Instant::now();
     inner_product_kernel2::launch(config, ctx, m, &d_a, &d_b, d_c_c, n)
         .expect("Kernel execution failed");
